requirements.txt
tdqm
phonenumbers
validate_email
py3dns